move `ComponentData` model to its own module so it can be reused
