+++ cyclonedx-bom/src/models/component.rs
@@ -590,14 +590,16 @@ mod test {
             attachment::Attachment,
             bom::BomReference,
             code::{Commit, Patch, PatchClassification},
+            component_data::{
+                ComponentData, ComponentDataType, DataContents, Graphic, GraphicsCollection,
+            },
             data_governance::{DataGovernance, DataGovernanceResponsibleParty},
             external_reference::{ExternalReference, ExternalReferenceType, Uri},
             hash::{Hash, HashAlgorithm, HashValue},
             license::LicenseChoice,
             modelcard::{
-                ApproachType, ComponentData, ComponentDataType, ConfidenceInterval, Considerations,
-                DataContents, Dataset, Datasets, Graphic, GraphicsCollection, Inputs, MLParameter,
-                ModelParameters, ModelParametersApproach, Outputs, PerformanceMetric,
+                ApproachType, ConfidenceInterval, Considerations, Dataset, Datasets, Inputs,
+                MLParameter, ModelParameters, ModelParametersApproach, Outputs, PerformanceMetric,
                 PerformanceMetrics, QuantitativeAnalysis,
             },
             organization::OrganizationalContact,
